… go.sum
… vendor/github.com/cloudevents/sdk-go/protocol/pubsub/v2/write_pubsub_message.go
… vendor/modules.txt