## Proposed Changes

- Adds a delivery benchmark which tests publishing events to a retry queue
- Updates the pubsub protocol to include https://github.com/cloudevents/sdk-go/pull/565 which removes the data copy when writing pubsub messages.

## Benchmark Results
### Fanout Delivery
#### Time/Op
> name                                         old time/op    new time/op    delta
> DeliveryNoReply/0_bytes-16                     9.01µs ± 1%    9.00µs ± 1%     ~     (p=0.672 n=29+28)
> DeliveryNoReply/1000_bytes-16                  9.25µs ± 1%    9.27µs ± 1%   +0.25%  (p=0.030 n=29+30)
> DeliveryNoReply/1000000_bytes-16                146µs ± 2%     145µs ± 2%     ~     (p=0.113 n=25+27)
> DeliveryWithReply/0_bytes-16                   18.5µs ± 1%    18.5µs ± 1%     ~     (p=0.982 n=30+30)
> DeliveryWithReply/1000_bytes-16                31.0µs ± 1%    31.0µs ± 1%     ~     (p=0.635 n=29+29)
> DeliveryWithReply/1000000_bytes-16              375µs ± 1%     374µs ± 2%     ~     (p=0.701 n=29+30)
> DeliveryNoReplyFakeClient/0_bytes-16           2.49µs ± 2%    2.47µs ± 2%   -0.57%  (p=0.015 n=30+30)
> DeliveryNoReplyFakeClient/1000_bytes-16        2.50µs ± 2%    2.49µs ± 1%   -0.62%  (p=0.007 n=30+28)
> DeliveryNoReplyFakeClient/100000_bytes-16      2.51µs ± 1%    2.48µs ± 1%   -1.32%  (p=0.000 n=30+28)
> DeliveryWithReplyFakeClient/0_bytes-16         2.67µs ± 3%    2.66µs ± 3%     ~     (p=0.755 n=27+30)
> DeliveryWithReplyFakeClient/1000_bytes-16      2.66µs ± 3%    2.65µs ± 4%     ~     (p=0.133 n=30+30)
> DeliveryWithReplyFakeClient/100000_bytes-16    2.62µs ± 3%    2.63µs ± 3%     ~     (p=0.897 n=29+30)
> DeliveryWithRetry/0_bytes-16                   30.6µs ± 1%    30.6µs ± 1%     ~     (p=0.510 n=29+30)
> DeliveryWithRetry/1000_bytes-16                30.3µs ± 1%    30.4µs ± 1%     ~     (p=0.068 n=29+28)
> DeliveryWithRetry/1000000_bytes-16             1.01ms ± 3%    0.90ms ± 4%  -10.46%  (p=0.000 n=27+27)

#### Bytes/Op
> name                                         old alloc/op   new alloc/op   delta
> DeliveryNoReply/0_bytes-16                     7.39kB ± 0%    7.39kB ± 0%     ~     (p=0.594 n=27+29)
> DeliveryNoReply/1000_bytes-16                  7.71kB ± 0%    7.72kB ± 0%     ~     (p=0.289 n=30+30)
> DeliveryNoReply/1000000_bytes-16               7.76kB ± 0%    7.75kB ± 0%     ~     (p=0.343 n=30+30)
> DeliveryWithReply/0_bytes-16                   14.6kB ± 0%    14.6kB ± 0%     ~     (p=0.642 n=28+30)
> DeliveryWithReply/1000_bytes-16                49.5kB ± 0%    49.5kB ± 0%     ~     (p=0.225 n=30+30)
> DeliveryWithReply/1000000_bytes-16             49.6kB ± 0%    49.6kB ± 0%     ~     (p=0.061 n=30+30)
> DeliveryNoReplyFakeClient/0_bytes-16           3.30kB ± 0%    3.30kB ± 0%     ~     (all equal)
> DeliveryNoReplyFakeClient/1000_bytes-16        3.39kB ± 0%    3.39kB ± 0%     ~     (all equal)
> DeliveryNoReplyFakeClient/100000_bytes-16      3.39kB ± 0%    3.39kB ± 0%     ~     (all equal)
> DeliveryWithReplyFakeClient/0_bytes-16         4.90kB ± 0%    4.90kB ± 0%     ~     (all equal)
> DeliveryWithReplyFakeClient/1000_bytes-16      4.99kB ± 0%    4.99kB ± 0%     ~     (all equal)
> DeliveryWithReplyFakeClient/100000_bytes-16    4.99kB ± 0%    4.99kB ± 0%     ~     (all equal)
> DeliveryWithRetry/0_bytes-16                   24.2kB ± 0%    24.2kB ± 0%     ~     (p=0.403 n=28+30)
> DeliveryWithRetry/1000_bytes-16                29.4kB ± 0%    28.3kB ± 0%   -3.66%  (p=0.000 n=30+30)
> DeliveryWithRetry/1000000_bytes-16             4.55MB ± 1%    3.55MB ± 1%  -22.10%  (p=0.000 n=30+30)

#### Allocs/Op
> name                                         old allocs/op  new allocs/op  delta
> DeliveryNoReply/0_bytes-16                        122 ± 0%       122 ± 0%     ~     (all equal)
> DeliveryNoReply/1000_bytes-16                     132 ± 0%       132 ± 0%     ~     (all equal)
> DeliveryNoReply/1000000_bytes-16                  132 ± 0%       132 ± 0%     ~     (all equal)
> DeliveryWithReply/0_bytes-16                      237 ± 0%       237 ± 0%     ~     (all equal)
> DeliveryWithReply/1000_bytes-16                   273 ± 0%       273 ± 0%     ~     (all equal)
> DeliveryWithReply/1000000_bytes-16                304 ± 0%       304 ± 0%     ~     (all equal)
> DeliveryNoReplyFakeClient/0_bytes-16             68.4 ± 1%      68.5 ± 1%     ~     (p=0.441 n=30+30)
> DeliveryNoReplyFakeClient/1000_bytes-16          71.3 ± 1%      71.6 ± 1%     ~     (p=0.073 n=30+30)
> DeliveryNoReplyFakeClient/100000_bytes-16        71.5 ± 1%      71.3 ± 1%     ~     (p=0.197 n=30+30)
> DeliveryWithReplyFakeClient/0_bytes-16           96.0 ± 0%      96.0 ± 0%     ~     (all equal)
> DeliveryWithReplyFakeClient/1000_bytes-16        99.0 ± 0%      99.0 ± 0%     ~     (all equal)
> DeliveryWithReplyFakeClient/100000_bytes-16      99.0 ± 0%      99.0 ± 0%     ~     (all equal)
> DeliveryWithRetry/0_bytes-16                      450 ± 0%       450 ± 0%     ~     (all equal)
> DeliveryWithRetry/1000_bytes-16                   464 ± 0%       462 ± 0%   -0.43%  (p=0.000 n=30+30)
> DeliveryWithRetry/1000000_bytes-16                516 ± 5%       510 ± 4%     ~     (p=0.104 n=30+30)

### Ingress
#### Time/Op
> name                            old time/op    new time/op    delta
> IngressHandler/0_bytes-16         9.61µs ± 2%    9.67µs ± 2%     ~     (p=0.325 n=10+10)
> IngressHandler/1000_bytes-16      10.1µs ± 3%    10.0µs ± 2%     ~     (p=0.342 n=10+10)
> IngressHandler/100000_bytes-16     122µs ± 3%     115µs ±12%   -5.64%  (p=0.015 n=8+9)

#### Bytes/OP
> name                            old alloc/op   new alloc/op   delta
> IngressHandler/0_bytes-16         15.3kB ± 0%    15.3kB ± 0%   -0.10%  (p=0.014 n=9+10)
> IngressHandler/1000_bytes-16      22.3kB ± 0%    21.2kB ± 0%   -4.81%  (p=0.000 n=10+10)
> IngressHandler/100000_bytes-16     759kB ± 0%     653kB ± 0%  -13.96%  (p=0.000 n=8+10)

#### Allocs/OP
> name                            old allocs/op  new allocs/op  delta
> IngressHandler/0_bytes-16            269 ± 0%       269 ± 0%     ~     (all equal)
> IngressHandler/1000_bytes-16         278 ± 0%       276 ± 0%   -0.72%  (p=0.000 n=10+10)
> IngressHandler/100000_bytes-16       310 ± 0%       308 ± 1%   -0.61%  (p=0.000 n=9+9)